requirements.txt
python-fasthtml
sqlite-minutils